+init test for sadad

+++ database/migrations/2017_03_02_000000_alter_iranpayment_transactions_table.php
@@ -19,6 +19,7 @@ class AlterIranPaymentTransactionsTable extends Migration
 
 	public function down()
 	{
+		if (app('config')->get('app.env') === 'testing') return;
         Schema::table($this->getTable(), function ($table) {
             $table->dropColumn('full_name');
             $table->dropColumn('email');
